+++ scripts/tests/test_bert_embedding.py
@@ -18,7 +18,6 @@
 # under the License.
 
 """Test BERTEmbedding."""
-import subprocess
 import time
 
 import pytest
